finished 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -45,6 +45,18 @@ blogsRouter.post("/", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
+    const token = request.token;
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    const blogToRemove = await Blog.findById(request.params.id);
+
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+
+    const user = await User.findById(decodedToken.id);
+    if (blogToRemove.user.toString() !== user._id.toString()) {
+      return response.status(401).json({ error: "invalid token" });
+    }
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
   } catch (exception) {
